Remove React code and Incorporate Changes from Nathan
First brings this module up to the changes that Nathan has made on the
AGR_UI hub.  Next removes some of the react code from the module... though
the test env running here still uses react.  Much of the code in this
commit is hard coded... will need to remove in the next.
Z