Refactor wordpress components
This change refactors the wordpress-related components to draw a clear distinction
between the three type of wordpress data we interact with: pages, posts, and a list
posts. This is reflected in the actions, reducers, and selectors too. Data fetching
work has been moved out of components and into actions, inline with other data types.