Various ticket fixes (#67)
* fixes ids

* removed console log statements

* bupming version

* AGR-2239
The Allele ID's are actually in the correctly named field now.  Needed 
to reflect that in code.

* AGR-2243
Fixing the allele symbol hover.  The symbol was being interpereted as a 
html tag before.  Also added some parsing to correct single variants w/ 
mutiple alleles.

* AGR-2241
Added a getJBrowseLink in Trackservice to generate links.  Implemented 
in both isoformandvarianttrack and isoformtrack.

* added notes so we do not undo this later

* bumping version

Co-authored-by: Nathan Dunn <ndunn@me.com>