Agr 2216 (#72)
* Initial Commit for track data filter
Based on the variant filter this is the start of a gene level filter.  I 
can see needing a transcript level filter too but for now committing 
this...

* Transcript Filter
Adding trans filter.  This will need cleaned up however it now filters 
based on transcript.  The isoformFilter function is not needed anymore 
and can be removed.  Also sets MAX_ROWS to 30 if a filter is passed.

* Removing unused code.

* Code Cleanup
Removing unnecessary fx, console.logs, and bumping version to 0.3.11