HTP View (#75)
Adding support to pass in a single HTP variant and display in a minimal 
capacity. Adding the option htpVariant and a location like X:1234 will 
display that location as a snp.  No interaction or thumbnail is 
provided.  This only works with the ISOFORM track and it does not make 
an additional apollo call.