Scrum 1564 Adding Xenbase Logo (#1059)
* Adding Header and Members Icon
Added the header and member icon for xenbase.  Currently the members 
icon is not consistent with the others.  In addition if we want to be 
consistent the member icon should link to the alliance mod landing page 
and not back to the mod as is stated in the ticket.

* Changing Link
Link now goes to Xenbase homepage.  Note... this is different to the 
other icons that link to the alliance mod landing page.

* Changed Xenbase logo
Changed the Xenbase logo to the one provided by Howie to be the correct 
size.

* Alphabetical Order
Putting the icons in the correct order.  Hoping a new commit will 
trigger the checks to rerun.